+++ include/MusicInfo.hpp
@@ -5,12 +5,12 @@
 
 class MusicInfo {
 public:
-  MusicInfo(String, String, String, int, std::vector<int>);
+  MusicInfo(String, String, String, double, std::vector<int>);
 
   String getTitle() const { return title; };
   String getArtist() const { return artist; };
   String getBpm() const { return bpm; };
-  int getOffset() const { return offset; };
+  double getOffset() const { return offset; };
   std::vector<int> getPlayLevels() const { return playLevels; };
   
 
@@ -19,7 +19,7 @@ private:
   String artist;
   String bpm;
   std::vector<int> playLevels; 
-  int offset;
+  double offset;
 
 };
